src/load_real_data.py
"""Convert Kaggle insurance claims dataset into our app's format.

Source: https://www.kaggle.com/datasets/mexwell/insurance-claims
1000 real insurance claims with fraud labels, 40 columns.
"""

import json
import os
import random
import pandas as pd
from datetime import datetime, timedelta

RAW_PATH = os.path.join(os.path.dirname(__file__), "..", "data", "raw", "insurance_claims.csv")
OUTPUT_DIR = os.path.join(os.path.dirname(__file__), "..", "data")


def load_and_convert():
    df = pd.read_csv(RAW_PATH)
    # Drop the empty trailing column
    df = df.drop(columns=["_c39"], errors="ignore")

    applicants = []
    policies = []
    claims = []
    medical_records = []

    seen_applicants = {}

    for idx, row in df.iterrows():
        policy_num = str(row["policy_number"])

        # ── Applicant (deduplicate by policy_number as proxy for person) ──
        app_id = f"APP-{idx+1:04d}"
        seen_applicants[policy_num] = app_id

        # Map fraud_reported Y/N to numeric score
        fraud_score = 0
        if row["fraud_reported"] == "Y":
            fraud_score = random.randint(3, 5)
        else:
            fraud_score = random.randint(0, 2)

        # Map incident_severity to a numeric indicator
        severity_map = {
            "Trivial Damage": 1,
            "Minor Damage": 2,
            "Major Damage": 3,
            "Total Loss": 4,
        }
        severity_num = severity_map.get(row.get("incident_severity", ""), 2)

        # Derive smoker heuristic from hobbies (not real, but enriches demo)
        hobbies = str(row.get("insured_hobbies", ""))
        smoker = hobbies.lower() in ("sleeping", "exercise")  # arbitrary for demo

        applicants.append({
            "applicant_id": app_id,
            "name": f"Policyholder {policy_num}",
            "age": int(row["age"]),
            "gender": row["insured_sex"].title(),
            "occupation": str(row["insured_occupation"]).replace("-", " ").title(),
            "income": int(row.get("capital-gains", 0)) + random.randint(30000, 80000),
            "health_status": random.choice(["Excellent", "Good", "Fair"]) if severity_num < 3 else "Fair",
            "smoker": smoker,
            "bmi": round(random.uniform(19.0, 35.0), 1),
            "exercise_frequency": random.choice(["Never", "Rarely", "Weekly", "Daily"]),
            "family_history": {
                "heart_disease": random.choice([True, False]),
                "diabetes": random.choice([True, False]),
                "cancer": random.choice([True, False]),
            },
            "previous_claims": random.randint(0, 3),
            "credit_score": random.randint(550, 820),
            "address": f"{row.get('incident_location', '')}, {row.get('incident_city', '')}, {row.get('incident_state', '')}",
            "phone": f"({random.randint(200,999)}) {random.randint(200,999)}-{random.randint(1000,9999)}",
            "email": f"policy{policy_num}@example.com",
            "education": str(row.get("insured_education_level", "Unknown")),
            "hobbies": hobbies,
            "months_as_customer": int(row.get("months_as_customer", 0)),
            "insured_relationship": str(row.get("insured_relationship", "")),
            "created_date": str(row.get("policy_bind_date", "2020-01-01")),
        })

        # ── Policy ────────────────────────────────────────────────────
        pol_id = f"POL-{idx+1:04d}"
        bind_date = str(row.get("policy_bind_date", "2020-01-01"))

        # Determine policy type from incident type
        incident_type = str(row.get("incident_type", ""))
        if "Vehicle" in incident_type or "Collision" in incident_type:
            policy_type = "Auto"
        elif "Theft" in incident_type:
            policy_type = "Auto"
        elif "Parked" in incident_type:
            policy_type = "Auto"
        else:
            policy_type = "Auto"

        csl = str(row.get("policy_csl", "100/300"))
        csl_parts = csl.split("/")
        coverage = int(csl_parts[-1]) * 1000 if csl_parts else 300000

        policies.append({
            "policy_id": pol_id,
            "applicant_id": app_id,
            "policy_type": policy_type,
            "coverage_amount": coverage,
            "premium_monthly": round(float(row.get("policy_annual_premium", 1200)) / 12, 2),
            "premium_annual": float(row.get("policy_annual_premium", 1200)),
            "deductible": int(row.get("policy_deductable", 1000)),
            "start_date": bind_date,
            "end_date": str(
                (datetime.strptime(bind_date, "%Y-%m-%d") + timedelta(days=365)).date()
            ) if bind_date != "nan" else "2021-01-01",
            "status": "Active",
            "policy_state": str(row.get("policy_state", "")),
            "policy_csl": csl,
            "umbrella_limit": int(row.get("umbrella_limit", 0)),
        })

        # ── Claim ─────────────────────────────────────────────────────
        clm_id = f"CLM-{idx+1:04d}"
        claims.append({
            "claim_id": clm_id,
            "applicant_id": app_id,
            "policy_id": pol_id,
            "claim_type": incident_type,
            "claim_amount": int(row.get("total_claim_amount", 0)),
            "injury_claim": int(row.get("injury_claim", 0)),
            "property_claim": int(row.get("property_claim", 0)),
            "vehicle_claim": int(row.get("vehicle_claim", 0)),
            "policy_limit": coverage,
            "claim_date": str(row.get("incident_date", "2015-01-01")),
            "days_since_policy_start": 0,
            "status": "Under Investigation" if row["fraud_reported"] == "Y" else random.choice(["Approved", "Pending", "Under Review"]),
            "description": (
                f"{incident_type} incident in {row.get('incident_city', 'Unknown')}, {row.get('incident_state', '')}. "
                f"Severity: {row.get('incident_severity', 'Unknown')}. "
                f"Collision type: {row.get('collision_type', 'N/A')}. "
                f"Vehicles involved: {row.get('number_of_vehicles_involved', 1)}. "
                f"Bodily injuries: {row.get('bodily_injuries', 0)}. "
                f"Property damage: {row.get('property_damage', 'Unknown')}. "
                f"Authorities contacted: {row.get('authorities_contacted', 'None')}. "
                f"Vehicle: {row.get('auto_year', '')} {row.get('auto_make', '')} {row.get('auto_model', '')}."
            ),
            "fraud_indicators": fraud_score,
            "fraud_reported": str(row.get("fraud_reported", "N")),
            "incident_severity": str(row.get("incident_severity", "")),
            "collision_type": str(row.get("collision_type", "")),
            "authorities_contacted": str(row.get("authorities_contacted", "")),
            "incident_city": str(row.get("incident_city", "")),
            "incident_state": str(row.get("incident_state", "")),
            "incident_location": str(row.get("incident_location", "")),
            "incident_hour": int(row.get("incident_hour_of_the_day", 0)),
            "vehicles_involved": int(row.get("number_of_vehicles_involved", 1)),
            "bodily_injuries": int(row.get("bodily_injuries", 0)),
            "witnesses": int(row.get("witnesses", 0)),
            "police_report": str(row.get("police_report_available", "NO")) == "YES",
            "property_damage": str(row.get("property_damage", "NO")) == "YES",
            "auto_make": str(row.get("auto_make", "")),
            "auto_model": str(row.get("auto_model", "")),
            "auto_year": int(row.get("auto_year", 2010)),
            "damage_location": f"{row.get('incident_city', '')}, {row.get('incident_state', '')}",
        })

        # ── Medical Record ────────────────────────────────────────────
        bodily_injuries = int(row.get("bodily_injuries", 0))
        conditions = []
        if bodily_injuries > 0:
            conditions.append("Trauma from accident")
        if severity_num >= 3:
            conditions.append("Severe impact injuries")
        if random.random() > 0.6:
            conditions.append(random.choice(["Hypertension", "Type 2 Diabetes", "Asthma", "Arthritis"]))

        medical_records.append({
            "applicant_id": app_id,
            "last_checkup": str(row.get("incident_date", "2015-01-01")),
            "blood_pressure": f"{random.randint(110, 160)}/{random.randint(65, 95)}",
            "cholesterol": random.randint(160, 280),
            "blood_sugar": random.randint(75, 180),
            "weight_lbs": random.randint(130, 280),
            "height_inches": random.randint(62, 76),
            "chronic_conditions": conditions,
            "medications": random.sample(["Lisinopril", "Metformin", "Ibuprofen", "Sertraline", "Atorvastatin"], random.randint(0, 2)),
            "allergies": random.sample(["Peanuts", "Penicillin", "None"], random.randint(0, 1)),
            "surgeries": ["Accident-related surgery"] if bodily_injuries >= 2 else [],
            "hospitalizations_last_5y": bodily_injuries,
            "smoker": smoker,
            "alcohol_use": random.choice(["None", "Social", "Moderate"]),
            "injury_from_claim": bodily_injuries > 0,
            "injury_count": bodily_injuries,
        })

    # ── Save ──────────────────────────────────────────────────────────
    os.makedirs(OUTPUT_DIR, exist_ok=True)
    for name, data in [
        ("applicants.json", applicants),
        ("policies.json", policies),
        ("claims.json", claims),
        ("medical_records.json", medical_records),
    ]:
        path = os.path.join(OUTPUT_DIR, name)
        with open(path, "w") as f:
            json.dump(data, f, indent=2)
        print(f"  Wrote {len(data)} records to {name}")

    # ── Summary Stats ─────────────────────────────────────────────────
    fraud_count = len([c for c in claims if c["fraud_reported"] == "Y"])
    total_exposure = sum(c["claim_amount"] for c in claims)

    print(f"\n{'='*60}")
    print(f"REAL DATA LOADED SUCCESSFULLY")
    print(f"{'='*60}")
    print(f"  Source: Kaggle insurance_claims.csv (1000 records)")
    print(f"  Applicants: {len(applicants)}")
    print(f"  Policies:   {len(policies)}")
    print(f"  Claims:     {len(claims)}")
    print(f"  Medical:    {len(medical_records)}")
    print(f"  Fraud cases: {fraud_count} ({fraud_count/len(claims)*100:.1f}%)")
    print(f"  Total exposure: ${total_exposure:,.0f}")
    print(f"  Avg claim: ${total_exposure/len(claims):,.0f}")
    print(f"{'='*60}")


if __name__ == "__main__":
    load_and_convert()
